test: Change behavior of temporary tlb files deletion
